调整控件四角不按比例缩放，取消四边中点缩放;增加四边连接线拖拽点;

+++ item/dragpoint.h
@@ -47,7 +47,7 @@ protected:
 private:
     QRectF boundRect;
 
-    int radius;
+    qreal radius;
 
     PointType pointType;
